Disallow reading from uninitialized arrays
This breaking change (probably the biggest breaking change in the
history of VeriFast) is necessary to make VeriFast sound with respect to
the fact that C compilers perform optimizations that assume that the
program does not read uninitialized memory. (See e.g.
<https://www.ralfj.de/blog/2019/07/14/uninit.html> and
<https://godbolt.org/z/dYq1Gc9G4>.)

This commit introduces, for each points-to predicate with parameter
types T* and T (representing an initialized object of type T), a
corresponding "underscore predicate" with parameter types T* and
option<T> (representing a possibly-uninitialized object of type T):

The points-to predicates and corresponding underscore predicates are:

boolean -> bool_
character -> char_
u_character -> uchar_
short_integer -> short_
u_short_integer -> ushort_
integer -> int_
u_integer -> uint_
llong_integer -> llong_
u_llong_integer -> ullong_
integer_ -> integer__ (generic integer type points-to predicate)
pointer -> pointer_
float_ -> float__
double_ -> double__
long_double -> long_double_

Similarly, for each array predicate with parameter types T*, int, and
list<T>, it introduces a corresponding underscore predicate with
parameter types T*, int, and list<option<T> >:

bools -> bools_
chars -> chars_
uchars -> uchars_
shorts -> shorts_
ushorts -> ushorts_
ints -> ints_
uints -> uints_
llongs -> llongs_
ullongs -> ullongs_
integers_ -> integers__ (generic integer array predicate)
pointers -> pointers_
floats -> floats_
doubles -> doubles_
long_doubles -> long_doubles_

Writing consumes an underscore chunk and produces an initialized chunk;
reading requires an initialized chunk. Allocation produces an underscore
chunk and deallocation consumes an underscore chunk.

The syntaxes *p |-> _ and p[..N] |-> _ are now interpreted as denoting
an underscore chunk.

TODO:
- Produce/consume underscore chunk on allocation/deallocation of
  primitive-typed objects.
- Struct field underscore predicates
- Update the VeriFast Tutorial
- Improve soundness regarding struct padding
- Make VeriFast sound with respect to pointer provenance and effective
  types

+++ examples/crypto_ccs/annotated_api/havege_util.c
@@ -7,7 +7,7 @@ void r_int(struct havege_state* state, int* i)
                [?f]havege_state_initialized(state) &*&
                random_permission(?principal, ?count) &*&
                havege_util(pub, ?proof_pred, principal) &*&
-               integer(i, _); @*/
+               int_(i, _); @*/
   /*@ ensures  [f]havege_state_initialized(state) &*&
                random_permission(principal, count + 1) &*&
                havege_util(pub, proof_pred, principal) &*&
@@ -17,6 +17,7 @@ void r_int(struct havege_state* state, int* i)
   //@ assert is_principal_with_public_nonces(?proof, pub, proof_pred, principal);
   char* temp = (void*) i;
   //@ close random_request(principal, 0, false);
+  //@ int__to_chars_(i);
   if (havege_random(state, temp, sizeof(int)) != 0) abort();
   //@ open cryptogram(temp, ?l, ?cs, ?random);
   /*@ if (!col)
